test(input): migrate legacy input-events
